Add stat_panel_counts() and expand tests
I added several tests for stat_quadrant_counts() and the new stat_panel_counts() but could increase the coverage very little and only by moving some tests from the compute function to the statistics. It looks like the code in the compute function is not seen as tested even if it is executed in the tests.