Fix text and label legends
This fix replaces the previous easy kludge that removed support for colour definitions containing an alpha value. This, different approach, supports colour definitions containing an alpha value, as `geom_text()` and `geom_label()` do. The default alpha and fill values in the geoms had to be changed to match those in 'ggplot2'.